Split wireless_communication_status_screen.c out
