+++ ld_script.txt
@@ -65,6 +65,7 @@ SECTIONS {
         src/union_room.o(.text);
         src/mystery_gift.o(.text);
         src/union_room_player_avatar.o(.text);
+        src/wireless_communication_status_screen.o(.text);
         src/union_room_battle.o(.text);
         src/mevent2.o(.text);
         src/mevent_801BAAC.o(.text);
@@ -450,6 +451,7 @@ SECTIONS {
         src/union_room.o(.rodata);
         src/mystery_gift.o(.rodata);
         src/union_room_player_avatar.o(.rodata);
+        src/wireless_communication_status_screen.o(.rodata);
         src/union_room_battle.o(.rodata);
         src/mevent2.o(.rodata);
         src/mevent_801BAAC.o(.rodata);
