+++ sym_bss.txt
@@ -10,6 +10,7 @@
 	.include "src/link_rfu_3.o"
 	.include "src/link_rfu_2.o"
 	.include "src/union_room.o"
+	.include "src/wireless_communication_status_screen.o"
 	.include "src/union_room_battle.o"
 	.include "src/dodrio_berry_picking.o"
 	.include "src/rtc.o"
